python-django/requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
packaging==25.0
pluggy==1.6.0
psycopg==3.2.6
psycopg-binary==3.2.6
Pygments==2.19.1
pytest==8.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
